+++ code/povelikin-rostislav/src/main/java/ru/unn/agile/redblacktree/RedBlackTree.java
@@ -1,4 +1,4 @@
-package ru.unn.agile.redblacktree.Model;
+package ru.unn.agile.redblacktree;
 
 public class RedBlackTree {
     private class Node {
